rename: CommonEditorComponent
