+++ src/components/common/Editor/CommonEditorComponent.tsx
